Remove legacy errors
