+++ query-engine/core-tests/tests/query_validation_tests/invalid_argument_type.expected.json
@@ -1,1 +1,1 @@
-{"is_panic":false,"message":"Failed to validate the query: `Unable to match input value to any allowed input type for the field. Parse errors: [Query parsing/validation error. Message: Invalid argument type. `id` should be of any of the following types: `IntFilter`, Query parsing/validation error. Message: Invalid argument type. `id` should be of any of the following types: `Int`]` at `Query.findManyUser.where.UserWhereInput.id`","meta":{"query_validation_error":"Unable to match input value to any allowed input type for the field. Parse errors: [Query parsing/validation error. Message: Invalid argument type. `id` should be of any of the following types: `IntFilter`, Query parsing/validation error. Message: Invalid argument type. `id` should be of any of the following types: `Int`]","query_position":"Query.findManyUser.where.UserWhereInput.id"},"error_code":"P2009"}
+{"is_panic":false,"message":"Unable to match input value to any allowed input type for the field. Parse errors: [Invalid argument type. `id` should be of any of the following types: `IntFilter`, Invalid argument type. `id` should be of any of the following types: `Int`]","meta":{"kind":"Union","errors":[{"kind":"InvalidArgumentType","argumentPath":["where","id"],"argument":{"name":"id","typeNames":["IntFilter"]},"selectionPath":["findManyUser"]},{"kind":"InvalidArgumentType","argumentPath":["where","id"],"argument":{"name":"id","typeNames":["Int"]},"selectionPath":["findManyUser"]}]},"error_code":"P2009"}
