+++ query-engine/core-tests/tests/query_validation_tests/invalid_argument_value.expected.json
@@ -1,1 +1,1 @@
-{"is_panic":false,"message":"Failed to validate the query: `Unable to match input value to any allowed input type for the field. Parse errors: [Query parsing/validation error at `Mutation.createOneUser.data.UserCreateInput.dob`: Unable to match input value to any allowed input type for the field. Parse errors: [Query parsing/validation error. Message: Invalid argument agument value. `1983-10-31R10:20:00.000Z` is not a valid `ISO-8601 DateTime`. Underlying error: input contains invalid characters, Query parsing/validation error. Message: Invalid argument type. `dob` should be of any of the following types: `Null`], Query parsing/validation error at `Mutation.createOneUser.data.UserUncheckedCreateInput.dob`: Unable to match input value to any allowed input type for the field. Parse errors: [Query parsing/validation error. Message: Invalid argument agument value. `1983-10-31R10:20:00.000Z` is not a valid `ISO-8601 DateTime`. Underlying error: input contains invalid characters, Query parsing/validation error. Message: Invalid argument type. `dob` should be of any of the following types: `Null`]]` at `Mutation.createOneUser.data`","meta":{"query_validation_error":"Unable to match input value to any allowed input type for the field. Parse errors: [Query parsing/validation error at `Mutation.createOneUser.data.UserCreateInput.dob`: Unable to match input value to any allowed input type for the field. Parse errors: [Query parsing/validation error. Message: Invalid argument agument value. `1983-10-31R10:20:00.000Z` is not a valid `ISO-8601 DateTime`. Underlying error: input contains invalid characters, Query parsing/validation error. Message: Invalid argument type. `dob` should be of any of the following types: `Null`], Query parsing/validation error at `Mutation.createOneUser.data.UserUncheckedCreateInput.dob`: Unable to match input value to any allowed input type for the field. Parse errors: [Query parsing/validation error. Message: Invalid argument agument value. `1983-10-31R10:20:00.000Z` is not a valid `ISO-8601 DateTime`. Underlying error: input contains invalid characters, Query parsing/validation error. Message: Invalid argument type. `dob` should be of any of the following types: `Null`]]","query_position":"Mutation.createOneUser.data"},"error_code":"P2009"}
+{"is_panic":false,"message":"Unable to match input value to any allowed input type for the field. Parse errors: [Unable to match input value to any allowed input type for the field. Parse errors: [Invalid argument agument value. `1983-10-31R10:20:00.000Z` is not a valid `ISO-8601 DateTime`. Underlying error: input contains invalid characters, Invalid argument type. `dob` should be of any of the following types: `Null`], Unable to match input value to any allowed input type for the field. Parse errors: [Invalid argument agument value. `1983-10-31R10:20:00.000Z` is not a valid `ISO-8601 DateTime`. Underlying error: input contains invalid characters, Invalid argument type. `dob` should be of any of the following types: `Null`]]","meta":{"kind":"Union","errors":[{"kind":"Union","errors":[{"kind":"InvalidArgumentValue","argumentPath":["data","dob"],"argument":{"name":"dob","typeNames":["ISO-8601 DateTime"]},"underlying_error":"input contains invalid characters"},{"kind":"InvalidArgumentType","argumentPath":["data","dob"],"argument":{"name":"dob","typeNames":["Null"]},"selectionPath":["createOneUser"]}]},{"kind":"Union","errors":[{"kind":"InvalidArgumentValue","argumentPath":["data","dob"],"argument":{"name":"dob","typeNames":["ISO-8601 DateTime"]},"underlying_error":"input contains invalid characters"},{"kind":"InvalidArgumentType","argumentPath":["data","dob"],"argument":{"name":"dob","typeNames":["Null"]},"selectionPath":["createOneUser"]}]}]},"error_code":"P2009"}
