+++ query-engine/core-tests/tests/query_validation_tests/unknown_input_field.expected.json
@@ -1,1 +1,1 @@
-{"is_panic":false,"message":"`Query.findManyUser.where.UserWhereInput.foo`: Field does not exist in enclosing type.","meta":{"kind":"UnknownInputField","inputType":{"name":"prisma.UserWhereInput","fields":[{"name":"AND","typeNames":["UserWhereInput","UserWhereInput[]"],"required":false},{"name":"OR","typeNames":["UserWhereInput[]"],"required":false},{"name":"NOT","typeNames":["UserWhereInput","UserWhereInput[]"],"required":false},{"name":"id","typeNames":["IntFilter","Int"],"required":false},{"name":"email","typeNames":["StringFilter","String"],"required":false},{"name":"dob","typeNames":["DateTimeNullableFilter","DateTime","Null"],"required":false}]},"selectionPath":["Query","findManyUser","where","UserWhereInput","foo"]},"error_code":"P2009"}
+{"is_panic":false,"message":"`findManyUser.foo`: Field does not exist in enclosing type.","meta":{"kind":"UnknownInputField","inputType":{"name":"prisma.UserWhereInput","fields":[{"name":"AND","typeNames":["UserWhereInput","UserWhereInput[]"],"required":false},{"name":"OR","typeNames":["UserWhereInput[]"],"required":false},{"name":"NOT","typeNames":["UserWhereInput","UserWhereInput[]"],"required":false},{"name":"id","typeNames":["IntFilter","Int"],"required":false},{"name":"email","typeNames":["StringFilter","String"],"required":false},{"name":"dob","typeNames":["DateTimeNullableFilter","DateTime","Null"],"required":false}]},"selectionPath":["findManyUser","foo"]},"error_code":"P2009"}
