+++ query-engine/core-tests/tests/query_validation_tests/value_too_large.expected.json
@@ -1,1 +1,1 @@
-{"is_panic":false,"message":"Failed to validate the query: `Unable to match input value to any allowed input type for the field. Parse errors: [Query parsing/validation error. Message: Invalid argument type. `id` should be of any of the following types: `IntFilter`, Query parsing/validation error. Message: Unable to fit float value (or large JS integer serialized in exponent notation) '18446744073709550000' into a 64 Bit signed integer for field 'id'. If you're trying to store large integers, consider using `BigInt`]` at `Query.findManyUser.where.UserWhereInput.id`","meta":{"query_validation_error":"Unable to match input value to any allowed input type for the field. Parse errors: [Query parsing/validation error. Message: Invalid argument type. `id` should be of any of the following types: `IntFilter`, Query parsing/validation error. Message: Unable to fit float value (or large JS integer serialized in exponent notation) '18446744073709550000' into a 64 Bit signed integer for field 'id'. If you're trying to store large integers, consider using `BigInt`]","query_position":"Query.findManyUser.where.UserWhereInput.id"},"error_code":"P2009"}
+{"is_panic":false,"message":"Unable to match input value to any allowed input type for the field. Parse errors: [Invalid argument type. `id` should be of any of the following types: `IntFilter`, Unable to fit float value (or large JS integer serialized in exponent notation) '18446744073709550000' into a 64 Bit signed integer for field 'id'. If you're trying to store large integers, consider using `BigInt`]","meta":{"kind":"Union","errors":[{"kind":"InvalidArgumentType","argumentPath":["where","id"],"argument":{"name":"id","typeNames":["IntFilter"]},"selectionPath":["findManyUser"]},{"kind":"ValueTooLarge","argumentPath":["where","id"],"argument":{"name":"id","typeNames":["BigInt"]},"selectionPath":["findManyUser"]}]},"error_code":"P2009"}
