+++ query-engine/core/src/query_graph_builder/builder.rs
@@ -40,9 +40,8 @@ impl QueryGraphBuilder {
     ) -> QueryGraphBuilderResult<(QueryGraph, IrSerializer)> {
         let mut selections = vec![selection];
         let mut parsed_object = QueryDocumentParser::new(crate::executor::get_request_now()).parse_object(
-            QueryPath::default(),
-            SelectionPath::default(),
-            ArgumentPath::default(),
+            Path::default(),
+            Path::default(),
             &selections,
             root_object,
         )?;
