+++ query-engine/core/src/query_graph_builder/error.rs
@@ -1,5 +1,6 @@
-use crate::{QueryGraphError, QueryParserError};
+use crate::QueryGraphError;
 use prisma_models::{DomainError, RelationFieldRef};
+use user_facing_errors::query_engine::validation::ValidationError;
 
 #[derive(Debug)]
 pub enum QueryGraphBuilderError {
@@ -19,7 +20,7 @@ pub enum QueryGraphBuilderError {
     },
 
     /// Wraps the initial parsing stage errors.
-    QueryParserError(QueryParserError),
+    QueryParserError(ValidationError),
 
     /// Wraps transformation errors from the prisma models.
     DomainError(DomainError),
@@ -75,8 +76,8 @@ impl From<DomainError> for QueryGraphBuilderError {
     }
 }
 
-impl From<QueryParserError> for QueryGraphBuilderError {
-    fn from(err: QueryParserError) -> Self {
+impl From<ValidationError> for QueryGraphBuilderError {
+    fn from(err: ValidationError) -> Self {
         QueryGraphBuilderError::QueryParserError(err)
     }
 }
